Use builder pattern for java process creation, allow caller to specify inheritIO (and warn when they do)
